Fix backup block deletion
We need to skip the backup to be deleted otherwise we would mark its
blocks as in use. Which would lead to no block deletion. We also need to
skip block deletion if we detect a in progress backup, since we cannot
determine yet which blocks the in progress backup will require.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>