Delete orphaned backup blocks
For the gc process we now find all referenced and orphaned backup blocks
afterwards we cleanup the orphaned blocks. We also now always update the
volume block count to the currently referenced and present blocks.

We are now also able to remove a corrupt backup file since we no longer
rely on the backup.cfg block mapping to find which blocks need to be
deleted

We try to limit block deletion to cases where we can evaluate whether it
is safe to do so. But there are still concurrency issues that can lead
to a race condition. When a new backup gets started right when we start
deleting backup blocks, and it just so happens that the new backup has a
block with the same checksum as one of the orphaned blocks that we are
going to delete. We do the below to evaluate whether it is safe to
cleanup the orphaned blocks:
- we make sure that we can find all blocks for a volume
- we make sure that we can load all backups
- we make sure that there are no in progress backups
- we make sure we can load the volume

Longhorn 1263 1435 1213

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>