Fix delete backup memory spike
Each volume backup is most likely to reference the same
block in the storage target. And having all backups with
block metadata and volume block metadata to check
reference count is memory-consuming.

Remove the use of the retained backup list and process a
single backup meta at a time when doing the reference
check.

Longhorn#2339

Signed-off-by: Chin-Ya Huang <chin-ya.huang@suse.com>