Fix CVE-2020-8912
Bump aws-sdk-go version to v1.34.2.

A vulnerability in the in-band key negotiation exists in the AWS S3 Crypto SDK for GoLang versions prior to V2. An attacker with write access to the targeted bucket can change the encryption algorithm of an object in the bucket, which can then allow them to change AES-GCM to AES-CTR. Using this in combination with a decryption oracle can reveal the authentication key used by AES-GCM as decrypting the GMAC tag leaves the authentication key recoverable as an algebraic equation. It is recommended to update your SDK to V2 or later, and re-encrypt your files. Affected versions < v1.34.2, the patched version v1.34.2.

We did not choose to bump to use https://github.com/aws/aws-sdk-go-v2/ because there is much change from https://github.com/aws/aws-sdk-go/ to https://github.com/aws/aws-sdk-go-v2/. But at least we bump the aws-sdk-go version to v1.34.2 to fix CVE-2020-8912.

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>